+++ v2/progon.h
@@ -1,5 +1,5 @@
 #ifndef _PROGON_H
 #define _PROGON_H 
 
-void matrix_solver(long double *A, long double *C, long double *B, int N, long double *F, long double *x);
+void matrix_solver(double *A, double *C, double *B, int N, double *F, double *x);
 #endif
