+++ v2/sheme.h
@@ -1,5 +1,5 @@
 #ifndef _SHEME_H_
 #define _SHEME_H_ 
-void solver(long double a, long double ymin, long double* x, long double dy, 
-	long double dt, int Nx, int Np, long double** gn, long double** g);
+void solver(double a, double ymin, double* x, double dy, 
+	double dt, int Nx, int Np, double** gn, double** g);
 #endif
